Updated for Lua 5.3, fixed cppcheck errors
- fixed issues reported by cppcheck
- updated for Lua 5.3
- ozCore
  * removed System::isInstrumented()

+++ src/common/luabase.hh
@@ -337,7 +337,7 @@
  * Shorthand for lua_tonumber
  */
 #define l_todouble(i) \
-  lua_tonumber(l, i)
+  double(lua_tonumber(l, i))
 
 /**
  * @def l_tostring
